Refactor DiamondGradingReport into separate modules/files in line with SDK naming conventions.
